# truncate user tasks