# Description
Introduce a feature to truncate or show all tasks.
Some accounts on dev have 100s of tasks and it can take a minute or two for the dashboard to load.
This is an attempt to load the dashboard faster.
With this change, the dashboard loads in about 15 seconds.

## Main changes explained:
If an account has more than four tasks, truncate it and show only three.
Add a button to show all tasks and to truncate the tasks.

## How to test:
1. Log into the dev branch and see which accounts have more than 4 tasks.
2. Log into this branch and check those accounts again.
3. Click "Show All (6) Tasks" to see all tasks.
4. Click "Truncate Tasks" to show only the first three tasks.

NOTE: Accounts with 100s of tasks are still slow.

## Screenshots or videos of changes:
| Truncate/Show SIX | Truncate/Show TWENTY |
|--------|--------|
| ![SIX](https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/97439fc6-1227-41fa-b62a-39e19f2bd8a7) | ![TWENTY](https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/1c3c6b4a-2213-46e6-b145-2fce4e82de63) | 

| LOAD BEFORE | LOAD AFTER |
|--------|--------|
| ![LOAD BEFORE](https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/6bdf2170-a633-4be9-a254-66cfbfab8c85) | ![LOAD AFTER](https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/e7d7d67e-0b20-45a9-a506-c690b93a6bc1) |

| 101 Tasks | 562 Tasks |
|--------|--------|
| <img width="617" alt="TASKS 101" src="https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/fa8a7126-5820-4b55-805c-f6fa64d3b268"> | <img width="620" alt="TASKS 562" src="https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/5bca900e-44db-4ec8-a925-e5394ab73d71"> |