@Cirakuze 2023-08-08
https://github.com/OneCommunityGlobal/HighestGoodNetworkApp/assets/15364639/71084a80-9abf-435a-97bb-5bceac731a1d

@one-community, updated to six tasks.